Merge all the code
Unfortunately to support both Android (using buildozer) and Linux the
best thing to do now is to merge all the package code.

I liked more when the project was splitted to multiple modules, but I
need the Android build and I want to be able to install on Linux using
pip.